flower api

+++ flowershop/flowershopbackend/FlowerShop.Domain/Entities/FlowerCategory.cs
@@ -9,7 +9,7 @@ namespace FlowerShop.Domain.Entities
 {
     public class FlowerCategory
     {
-        public int Id { get; set; }
+        public float Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public bool IsActive { get; set; }
